+++ frontend/src/pages/crawls.ts
@@ -9,8 +9,8 @@ import type { PageChangeEvent } from "@/components/ui/pagination";
 import { CrawlStatus } from "@/features/archived-items/crawl-status";
 import type { APIPaginatedList, APIPaginationQuery } from "@/types/api";
 import type { Crawl, CrawlState } from "@/types/crawler";
-import { needLogin } from "@/utils/auth";
 import { activeCrawlStates } from "@/utils/crawler";
+import needLogin from "@/utils/decorators/needLogin";
 import LiteElement, { html } from "@/utils/LiteElement";
 
 type SortField = "started" | "firstSeed" | "fileSize";
