+++ frontend/src/pages/home.ts
@@ -17,7 +17,9 @@ import { formatAPIUser } from "@/utils/user";
 
 /**
  * Home page when org is not selected.
- * Currently, only visible to superadmins--redirects to user's org, otherwise
+ * Currently, only visible to superadmins--redirects to user's org, otherwise.
+ * Uses custom redirect instead of needLogin decorator to suppress "need login"
+ * message when accessing root URL.
  *
  * TODO Refactor out superadmin UI
  */
