+++ frontend/src/pages/org/index.ts
@@ -19,8 +19,8 @@ import type { Crawl, JobType } from "@/types/crawler";
 import type { UserOrg } from "@/types/user";
 import { isApiError } from "@/utils/api";
 import type { ViewState } from "@/utils/APIRouter";
-import { needLogin } from "@/utils/auth";
 import { DEFAULT_MAX_SCALE } from "@/utils/crawler";
+import needLogin from "@/utils/decorators/needLogin";
 import LiteElement, { html } from "@/utils/LiteElement";
 import { isAdmin, isCrawler, type OrgData } from "@/utils/orgs";
 import { AppStateService } from "@/utils/state";
